Modify library to support different Clients.
- To allow UDP, zigbee,... clients, I have changed body of library. Now, there is a main class called MQTTSN. This class implements MQTTSN commands and functionality. MQTTSN needs a MqttsnClient as parameters.

MqttsnClient is the responsible of send and read MqttsnPackets depending on the link and physic layer used (Ethernet-UDP, zigbee,..)

+++ src/MQTTSNCommon.h
@@ -60,7 +60,7 @@ class MQTTSNCommon{
         void pingResp();
         void pubAck(uint16_t topicId, uint16_t msgId, uint8_t returnCode = ACCEPTED);
         bool subscribe(const char * topic, uint16_t * topicIdOut);
-        bool subscribe(unsigned int topic, uint16_t * topicIdOut);
+        bool subscribe(uint16_t topic, uint16_t * topicIdOut);
         bool unsubscribe(const char * topic);
         bool unsubscribe(uint16_t topic);
